Add 'boolean_operator' test in main.rs
A new test function 'boolean_operator' has been added to main.rs. The function tests the boolean operations using the variables 'absen' and 'nilai_akhir'. It evaluates if the specified conditions are met and outputs if these conditions resulted in passing the test. Thus, improving the test coverage for boolean operations.