Add Apple struct with addition functionality and tests
Added an Apple struct to src/main.rs with an implementation of the Add trait, facilitating the addition of quantities of Apples. Some basic tests have also been included to verify this functionality. In addition, the `double` function was created, this function doubles the value of an Option<i32> if it's not None. Tests for the 'double' function were also added.